% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bladder.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Bladder dataset}
\format{A data.frame with 191 rows and the following variables:
\describe{
\item{id}{}
\item{event}{}
\item{interval}{}
\item{inttime}{}
\item{start}{}
\item{stop}{}
\item{tx}{}
\item{num}{}
\item{size}{}
}}
\source{
Kleinbaum, D.G. and Klein, M. (2012) Survival Analysis - A
    Self-Learning Text, 3rd ed., Springer
}
\usage{
bladder
}
\description{
Bladder dataset.
}
\examples{
head(bladder, n = 10)
summary(bladder)
}
\keyword{datasets}

