% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evans.R
\docType{data}
\name{evans}
\alias{evans}
\title{Evans County dataset}
\format{
A data.frame with 50 rows and the following variables:
\describe{
\item{chr}{1 = history of chronic disease, 0 = no history of chronic disease}
\item{time}{Survival time (years)}
\item{event}{1 = patient diagnosed with coronary heart disease, 0 = censored}
}
}
\source{
Kleinbaum, D.G. and Klein, M. (2012) Survival Analysis - A
    Self-Learning Text, 3rd ed., Springer
}
\usage{
evans
}
\description{
A sample of 50 patients from the Evans County Heart Study dataset.
}
\details{
This dataset is used in the practice questions for Chapter 2.
}
\keyword{datasets}
