% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suanselete3-package.R
\docType{package}
\name{suanselete3}
\alias{suanselete3}
\alias{suanselete3-package}
\title{Dataset from Kleinbaum and Klein
"Survival analysis - A self learning text" (3rd Ed., 2012)}
\source{
Kleinbaum, D.G. and Klein, M. (2012) Survival Analysis - A
    Self-Learning Text, 3rd ed., Springer
}
\description{
Dataset from Kleinbaum and Klein
"Survival analysis - A self learning text" (3rd Ed., 2012)
}

