% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bladder.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Bladder dataset}
\format{
A data.frame with 191 rows and the following variables:
\describe{
\item{id}{Patient ID}
\item{event}{Patient had tumor (1) or not (0)}
\item{interval}{Which time interval it is for a given patient, where 1 is the first interval, 2 is the second, etc}
\item{start}{Starting time (in months) for the interval}
\item{stop}{Time of event or censorship for each interval}
\item{tx}{Treatment status (1 = thiotepa, 0 = placebo)}
\item{num}{Initial number of tumors}
\item{size}{Initial size (cm) of tumor}
}
}
\source{
Kleinbaum, D.G. and Klein, M. (2012) Survival Analysis - A
    Self-Learning Text, 3rd ed., Springer
}
\usage{
bladder
}
\description{
Interval data for bladder cancer patients after TURBT, with patients who
received placebo or thiotepa.
}
\keyword{datasets}
